pyproject.toml
[project]
name = "figma-helper"
version = "0.1.0"
description = ""
authors = [
    {name = "Brilev Denis",email = "brilevdv@gazdv.ru"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-telegram-bot[job-queue] (>=22.0,<23.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
